Time: 0 ms (100%), Space: 17.9 MB (23.21%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.cpp
@@ -30,7 +30,7 @@ public:
            
         }
 
-        return ans;
+        return (int)ans;
        
     }
 };
